A lot of corrections for x64 ( thx RBDOOM3 ), warning fixes, ignores

+++ neo/d3xp/physics/Physics_Player.cpp
@@ -691,7 +691,7 @@ void idPhysics_Player::WalkMove() {
 	if ( waterLevel ) {
 		float	waterScale;
 
-		waterScale = waterLevel / 3.0f;
+		waterScale = (float)waterLevel / 3.0f;
 		waterScale = 1.0f - ( 1.0f - PM_SWIMSCALE ) * waterScale;
 		if ( wishspeed > playerSpeed * waterScale ) {
 			wishspeed = playerSpeed * waterScale;
